// src/models/ResumeModel.ts
export interface ResumeModel {
  profileId: number;
  firstName: string;
  lastName: string;
  employmentMode: string;
  employmentType: string;
  location: string;
  totalExpInYrs: number;
  totalExpInMonths: number;
  skills: string;
  currentCTC: string;
  expectedCTC: string;
  isSaved: boolean;
}